Extract menu fucntion to menu module
